Split combat code from entities
